Move Thrift metastore authentication classes
