+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/thrift/MetastoreKerberosConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.authentication;
+package io.trino.plugin.hive.metastore.thrift;
 
 import io.airlift.configuration.Config;
 import io.airlift.configuration.ConfigDescription;
