+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/metastore/thrift/TestMetastoreKerberosConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.authentication;
+package io.trino.plugin.hive.metastore.thrift;
 
 import com.google.common.collect.ImmutableMap;
 import org.testng.annotations.Test;
